add mail route (#83)
I want @DinDin985 & @BBruington to use this PR as a building block for how future front-end features should be structured.  Notice the use of the _components folder.   

I also want to highlight the use of jotai for data, eventually this should come from the back end but for now this is a good middle ground. 




![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/j6fxR9uRkNh0AqyQ0whr/4c25bcbc-a1de-4aae-90db-e4b75d2a1f26.png)